Everything with a partial update is now a long double

+++ Data.cpp
@@ -84,9 +84,9 @@ void Data::compute_ray_grid()
 {
 	// Make vectors of the correct size
 	x_rays.assign(ni*resolution,
-			vector<double>(nj*resolution));
+			vector<long double>(nj*resolution));
 	y_rays.assign(ni*resolution,
-			vector<double>(nj*resolution));
+			vector<long double>(nj*resolution));
 
 	// Distance between adjacent rays
 	double L = dx/resolution;
