+++ MyModel.h
@@ -20,11 +20,11 @@ class MyModel:public DNest3::Model
 		double sigma0, sigma1;
 
 		// Source plane position of rays
-		std::vector< std::vector<double> > xs;
-		std::vector< std::vector<double> > ys;
+		std::vector< std::vector<long double> > xs;
+		std::vector< std::vector<long double> > ys;
 
 		// Surface brightness of rays
-		std::vector< std::vector<double> > surface_brightness;
+		std::vector< std::vector<long double> > surface_brightness;
 
 		// Model image
 		std::vector< std::vector<long double> > model_image;
